// src/types/route.ts
export type Route =
  | 'login'
  | 'register'
  | 'forgot-password'
  | 'reset-password'
  | 'chatbot'
  | 'about'
  | 'courses'
  | 'admission'
  | 'faq'
  | 'contact'
  | 'profile'
  | 'admin-login'
  | 'admin-dashboard'
  | 'admin-announcements'
  | 'admin-college-info'
  | 'admin-courses'
  | 'admin-faqs'
  | 'admin-knowledge'
  | 'admin-chat-history'
  | 'admin-settings';